feat(backend-refactor): models are now called entities
