+++ backend/src/main/java/com/trocandgo/trocandgo/DefaultDataInitializer.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.trocandgo.trocandgo.model.ServiceCategories;
+import com.trocandgo.trocandgo.entity.ServiceCategories;
 import com.trocandgo.trocandgo.repository.ServiceCategoryRepository;
 
 import jakarta.annotation.PostConstruct;
