+++ backend/src/main/java/com/trocandgo/trocandgo/controller/AuthControllerV1.java
@@ -20,9 +20,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.trocandgo.trocandgo.dto.request.LoginRequest;
 import com.trocandgo.trocandgo.dto.request.SignupRequest;
-import com.trocandgo.trocandgo.model.Roles;
-import com.trocandgo.trocandgo.model.Users;
-import com.trocandgo.trocandgo.model.enums.RoleName;
+import com.trocandgo.trocandgo.entity.Roles;
+import com.trocandgo.trocandgo.entity.Users;
+import com.trocandgo.trocandgo.entity.enums.RoleName;
 import com.trocandgo.trocandgo.repository.UserRepository;
 import com.trocandgo.trocandgo.service.JwtService;
 import com.trocandgo.trocandgo.service.UserDetailsImpl;
