+++ backend/src/main/java/com/trocandgo/trocandgo/dto/request/CreateServiceRequest.java
@@ -1,7 +1,7 @@
 package com.trocandgo.trocandgo.dto.request;
 
-import com.trocandgo.trocandgo.model.enums.ServiceStatusTitle;
-import com.trocandgo.trocandgo.model.enums.ServiceTypeTitle;
+import com.trocandgo.trocandgo.entity.enums.ServiceStatusTitle;
+import com.trocandgo.trocandgo.entity.enums.ServiceTypeTitle;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Value;
