+++ backend/src/main/java/com/trocandgo/trocandgo/entity/Favorites.java
@@ -1,4 +1,4 @@
-package com.trocandgo.trocandgo.model;
+package com.trocandgo.trocandgo.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
