+++ backend/src/main/java/com/trocandgo/trocandgo/entity/ServiceCategories.java
@@ -1,10 +1,7 @@
-package com.trocandgo.trocandgo.model;
+package com.trocandgo.trocandgo.entity;
 
-import com.trocandgo.trocandgo.model.enums.ServiceStatusTitle;
-
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -12,17 +9,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Getter
+@Setter
 @NoArgsConstructor
 @RequiredArgsConstructor
-public class ServiceStatuses {
+public class ServiceCategories {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
     @NonNull
-    @Enumerated(EnumType.STRING)
-    private ServiceStatusTitle title;
+    @Column(length = 255)
+    private String title;
 }
