+++ backend/src/main/java/com/trocandgo/trocandgo/entity/ServiceStatuses.java
@@ -1,6 +1,6 @@
-package com.trocandgo.trocandgo.model;
+package com.trocandgo.trocandgo.entity;
 
-import com.trocandgo.trocandgo.model.enums.RoleName;
+import com.trocandgo.trocandgo.entity.enums.ServiceStatusTitle;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -12,19 +12,17 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import lombok.ToString;
 
 @Entity
 @Getter
 @NoArgsConstructor
 @RequiredArgsConstructor
-@ToString
-public class Roles {
-   @Id
-   @GeneratedValue(strategy = GenerationType.IDENTITY)
-   private long id;
+public class ServiceStatuses {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
 
-   @NonNull
-   @Enumerated(EnumType.STRING)
-   private RoleName name;
+    @NonNull
+    @Enumerated(EnumType.STRING)
+    private ServiceStatusTitle title;
 }
