+++ backend/src/main/java/com/trocandgo/trocandgo/entity/ServiceTypes.java
@@ -1,7 +1,10 @@
-package com.trocandgo.trocandgo.model;
+package com.trocandgo.trocandgo.entity;
 
-import jakarta.persistence.Column;
+import com.trocandgo.trocandgo.entity.enums.ServiceTypeTitle;
+
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -9,19 +12,17 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor
 @RequiredArgsConstructor
-public class ServiceCategories {
+public class ServiceTypes {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
     @NonNull
-    @Column(length = 255)
-    private String title;
+    @Enumerated(EnumType.STRING)
+    private ServiceTypeTitle title;
 }
