+++ backend/src/main/java/com/trocandgo/trocandgo/entity/UnreadMessages.java
@@ -1,4 +1,4 @@
-package com.trocandgo.trocandgo.model;
+package com.trocandgo.trocandgo.entity;
 
 import java.util.Date;
