+++ backend/src/main/java/com/trocandgo/trocandgo/entity/Users.java
@@ -1,4 +1,4 @@
-package com.trocandgo.trocandgo.model;
+package com.trocandgo.trocandgo.entity;
 
 import java.util.HashSet;
 import java.util.Set;
