+++ backend/src/main/java/com/trocandgo/trocandgo/entity/enums/RoleName.java
@@ -1,4 +1,4 @@
-package com.trocandgo.trocandgo.model.enums;
+package com.trocandgo.trocandgo.entity.enums;
 
 public enum RoleName {
     ROLE_USER,
