+++ backend/src/main/java/com/trocandgo/trocandgo/repository/AddressRepository.java
@@ -2,6 +2,6 @@ package com.trocandgo.trocandgo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.trocandgo.trocandgo.model.Adresses;
+import com.trocandgo.trocandgo.entity.Adresses;
 
 public interface AddressRepository extends JpaRepository<Adresses, Long> {}
