+++ backend/src/main/java/com/trocandgo/trocandgo/repository/FavoriteRepository.java
@@ -2,10 +2,10 @@ package com.trocandgo.trocandgo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.trocandgo.trocandgo.model.Favorites;
-import com.trocandgo.trocandgo.model.FavoritesPK;
-import com.trocandgo.trocandgo.model.Services;
-import com.trocandgo.trocandgo.model.Users;
+import com.trocandgo.trocandgo.entity.Favorites;
+import com.trocandgo.trocandgo.entity.FavoritesPK;
+import com.trocandgo.trocandgo.entity.Services;
+import com.trocandgo.trocandgo.entity.Users;
 
 public interface FavoriteRepository extends JpaRepository<Favorites, FavoritesPK> {
     boolean existsByUserAndService(Users user, Services service);
