+++ backend/src/main/java/com/trocandgo/trocandgo/repository/ReviewRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.trocandgo.trocandgo.model.Reviews;
-import com.trocandgo.trocandgo.model.ReviewsPK;
-import com.trocandgo.trocandgo.model.Services;
+import com.trocandgo.trocandgo.entity.Reviews;
+import com.trocandgo.trocandgo.entity.ReviewsPK;
+import com.trocandgo.trocandgo.entity.Services;
 
 public interface ReviewRepository extends JpaRepository<Reviews, ReviewsPK> {
     Page<Reviews> findAllByService(Services service, Pageable pageable);
