+++ backend/src/main/java/com/trocandgo/trocandgo/repository/RoleRepository.java
@@ -2,8 +2,8 @@ package com.trocandgo.trocandgo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.trocandgo.trocandgo.model.Roles;
-import com.trocandgo.trocandgo.model.enums.RoleName;
+import com.trocandgo.trocandgo.entity.Roles;
+import com.trocandgo.trocandgo.entity.enums.RoleName;
 
 public interface RoleRepository extends JpaRepository<Roles, Long> {
     Roles getByName(RoleName name);
