+++ backend/src/main/java/com/trocandgo/trocandgo/repository/ServiceCategoryRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trocandgo.trocandgo.model.ServiceCategories;
+import com.trocandgo.trocandgo.entity.ServiceCategories;
 
 @Repository
 public interface ServiceCategoryRepository extends JpaRepository<ServiceCategories, Long> {
