+++ backend/src/main/java/com/trocandgo/trocandgo/repository/ServiceRepository.java
@@ -4,8 +4,8 @@ import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import com.trocandgo.trocandgo.model.Adresses;
-import com.trocandgo.trocandgo.model.Services;
+import com.trocandgo.trocandgo.entity.Adresses;
+import com.trocandgo.trocandgo.entity.Services;
 
 public interface ServiceRepository extends JpaRepository<Services, UUID>, JpaSpecificationExecutor<Services> {
     boolean existsByAdress(Adresses adress);
