+++ backend/src/main/java/com/trocandgo/trocandgo/repository/ServiceStatusRepository.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trocandgo.trocandgo.model.ServiceStatuses;
-import com.trocandgo.trocandgo.model.enums.ServiceStatusTitle;
+import com.trocandgo.trocandgo.entity.ServiceStatuses;
+import com.trocandgo.trocandgo.entity.enums.ServiceStatusTitle;
 
 @Repository
 public interface ServiceStatusRepository extends JpaRepository<ServiceStatuses, Long> {
