+++ backend/src/main/java/com/trocandgo/trocandgo/repository/ServiceTypeRepository.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trocandgo.trocandgo.model.ServiceTypes;
-import com.trocandgo.trocandgo.model.enums.ServiceTypeTitle;
+import com.trocandgo.trocandgo.entity.ServiceTypes;
+import com.trocandgo.trocandgo.entity.enums.ServiceTypeTitle;
 
 @Repository
 public interface ServiceTypeRepository extends JpaRepository<ServiceTypes, Long> {
