+++ backend/src/main/java/com/trocandgo/trocandgo/repository/UserRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.trocandgo.trocandgo.model.Users;
+import com.trocandgo.trocandgo.entity.Users;
 
 @Repository
 public interface UserRepository extends JpaRepository<Users, Long> {
