+++ backend/src/main/java/com/trocandgo/trocandgo/service/ServiceService.java
@@ -10,6 +10,11 @@ import org.springframework.stereotype.Service;
 import com.trocandgo.trocandgo.dto.request.CreateReviewRequest;
 import com.trocandgo.trocandgo.dto.request.CreateServiceRequest;
 import com.trocandgo.trocandgo.dto.request.SearchRequest;
+import com.trocandgo.trocandgo.entity.Reviews;
+import com.trocandgo.trocandgo.entity.ReviewsPK;
+import com.trocandgo.trocandgo.entity.ServiceStatuses;
+import com.trocandgo.trocandgo.entity.ServiceTypes;
+import com.trocandgo.trocandgo.entity.Services;
 import com.trocandgo.trocandgo.exception.InvalidPageableException;
 import com.trocandgo.trocandgo.exception.InvalidServiceCategoryException;
 import com.trocandgo.trocandgo.exception.InvalidUUIDException;
@@ -17,11 +22,6 @@ import com.trocandgo.trocandgo.exception.NotAuthenticatedException;
 import com.trocandgo.trocandgo.exception.ReviewAlreadyExistsException;
 import com.trocandgo.trocandgo.exception.SelfReviewException;
 import com.trocandgo.trocandgo.exception.ServiceNotFoundException;
-import com.trocandgo.trocandgo.model.Reviews;
-import com.trocandgo.trocandgo.model.ReviewsPK;
-import com.trocandgo.trocandgo.model.ServiceStatuses;
-import com.trocandgo.trocandgo.model.ServiceTypes;
-import com.trocandgo.trocandgo.model.Services;
 import com.trocandgo.trocandgo.repository.ReviewRepository;
 import com.trocandgo.trocandgo.repository.ServiceCategoryRepository;
 import com.trocandgo.trocandgo.repository.ServiceRepository;
