+++ backend/src/main/java/com/trocandgo/trocandgo/service/UserDetailsImpl.java
@@ -7,7 +7,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import com.trocandgo.trocandgo.model.Users;
+import com.trocandgo.trocandgo.entity.Users;
 
 import lombok.Getter;
