+++ backend/src/main/java/com/trocandgo/trocandgo/service/UserDetailsServiceImpl.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.trocandgo.trocandgo.model.Users;
+import com.trocandgo.trocandgo.entity.Users;
 import com.trocandgo.trocandgo.repository.UserRepository;
 
 @Service
